chore(2024/06): set up kotlin project
