+++ solutions/2024/day06/kotlin/.idea/.name
@@ -1,0 +1,1 @@
+Advent of Code [2024](day 6)
