+++ solutions/2024/day06/kotlin/.idea/ktfmt.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="KtfmtSettings">
+    <option name="enableKtfmt" value="Enabled" />
+    <option name="uiFormatterStyle" value="Kotlinlang" />
+  </component>
+</project>
