+++ solutions/2024/day06/kotlin/detekt.yml
@@ -1,0 +1,3 @@
+config:
+  validation: true
+  warningsAsErrors: false
