+++ solutions/2024/day06/kotlin/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Dec 01 11:54:58 CET 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.10-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
