+++ solutions/2024/day06/kotlin/settings.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
+}
+rootProject.name = "Advent of Code [2024](day 6)"
