+++ solutions/2024/day06/kotlin/src/main/kotlin/se/fzy/adventofcode/Resources.kt
@@ -1,0 +1,5 @@
+package se.fzy.adventofcode
+
+object Resources {
+    fun readInput(fileName: String): String = javaClass.getResource("/$fileName").readText()
+}
