+++ solutions/2024/day06/kotlin/src/main/resources/.gitignore
@@ -1,0 +1,1 @@
+*.txt
