+++ solutions/2024/day06/kotlin/src/test/kotlin/se/fzy/adventofcode/ExampleTest.kt
@@ -1,0 +1,12 @@
+package se.fzy.adventofcode
+
+import org.junit.jupiter.api.Test
+import strikt.api.expectThat
+import strikt.assertions.isTrue
+
+class ExampleTest {
+    @Test
+    fun `An example test`() {
+        expectThat(true).isTrue()
+    }
+}
